+++ library/src/com/jakewharton/android/view/CirclePageIndicator.java
@@ -14,7 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.jakewharton.android.viewpagerindicator;
+package com.jakewharton.android.view;
+
+import com.jakewharton.android.viewpagerindicator.R;
 
 import android.content.Context;
 import android.content.res.Resources;