+++ sample/src/com/jakewharton/android/viewpagerindicator/sample/SampleCirclesStyledTheme.java
@@ -3,7 +3,7 @@ package com.jakewharton.android.viewpagerindicator.sample;
 import android.os.Bundle;
 import android.support.v4.app.FragmentActivity;
 import android.support.v4.view.ViewPager;
-import com.jakewharton.android.viewpagerindicator.CirclePageIndicator;
+import com.jakewharton.android.view.CirclePageIndicator;
 
 public class SampleCirclesStyledTheme extends FragmentActivity {
 	@Override