+++ sample/src/com/jakewharton/android/viewpagerindicator/sample/SampleCirclesWithListener.java
@@ -4,7 +4,7 @@ import android.os.Bundle;
 import android.support.v4.app.FragmentActivity;
 import android.support.v4.view.ViewPager;
 import android.widget.Toast;
-import com.jakewharton.android.viewpagerindicator.CirclePageIndicator;
+import com.jakewharton.android.view.CirclePageIndicator;
 
 public class SampleCirclesWithListener extends FragmentActivity {
 	@Override