+++ sample/src/com/jakewharton/android/viewpagerindicator/sample/SampleTitlesStyledMethods.java
@@ -3,8 +3,8 @@ package com.jakewharton.android.viewpagerindicator.sample;
 import android.os.Bundle;
 import android.support.v4.app.FragmentActivity;
 import android.support.v4.view.ViewPager;
-import com.jakewharton.android.viewpagerindicator.TitlePageIndicator;
-import com.jakewharton.android.viewpagerindicator.TitlePageIndicator.IndicatorStyle;
+import com.jakewharton.android.view.TitlePageIndicator;
+import com.jakewharton.android.view.TitlePageIndicator.IndicatorStyle;
 
 public class SampleTitlesStyledMethods extends FragmentActivity {
 	@Override