+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Change Log
 ==========
 
+Version 1.1.0 *(In Development)*
+--------------------------------
+
+ * Package changed from `com.jakewharton.android.viewpagerindicator` to
+   `com.jakewharton.android.view`.
+
+
 Version 1.0.1 *(2011-09-15)*
 ----------------------------